// packages/webgpu-renderers/src/text-animations/types.ts
export const SelectorShapeSchema = z.enum([
	"square",
	"ramp_up",
	"ramp_down",
	"triangle",
	"round",
	"smooth",
]);
export type SelectorShape = z.infer<typeof SelectorShapeSchema>;

export const SelectorModeSchema = z.enum([
	"add",
	"subtract",
	"intersect",
	"min",
	"max",
	"difference",
]);
export type SelectorMode = z.infer<typeof SelectorModeSchema>;